Add Bit::Mask structure
Along with adding the bit::Mask structure, the codebase has been
refactored to use static constexpr bit::Mask objects when manipulating
bits using bit::Insert or bit::Extract

Updated .clang-format to better align function call parameters to make
the use of Bit::Maks structures, and the rest of the code, more readable.

+++ library/L1_Peripheral/system_controller.hpp
@@ -13,7 +13,7 @@ class SystemController
   class PeripheralID
   {
    public:
-    size_t device_id = -1;
+    uint8_t device_id = -1;
   };
   template <size_t kDeviceId>
   class AddPeripheralID : public PeripheralID
